Criando a camada de Infraestrutura

+++ TesteTecnico/Domain/Entities/Order.cs
@@ -9,27 +9,27 @@ namespace TT.Infra.Domain.Entities
     {
         public Customer Customer { get; private set; }
         public int CustomerId { get; private set; }
-        public float Price { get; private set; }
+        public decimal Price { get; private set; }
         public DateTime CreatedAt { get; private set; }
 
-        public Order(float price)
+        public Order(decimal price)
         {
             ValidateDomain(price);
         }
 
-        public Order(int id, float price)
+        public Order(int id, decimal price)
         {
             DomainExceptionValidation.When(id <= 0, "Invalid Id");
             Id = id;
             ValidateDomain(price);
         }
 
-        public void Update(float price)
+        public void Update(decimal price)
         {
             ValidateDomain(price);
         }
 
-        private void ValidateDomain(float price)
+        private void ValidateDomain(decimal price)
         {
             DomainExceptionValidation.When(price <= 0, "Invalid price. Must be greater than 0");
